fix springboot, add thymeleaf support, change jsp tags to thymeleaf
