+++ src/main/java/file/config/ThymeleafConfig.java
@@ -1,34 +1,15 @@
-package file.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.thymeleaf.spring5.SpringTemplateEngine;
-import org.thymeleaf.spring5.view.ThymeleafViewResolver;
-import org.thymeleaf.templateresolver.ServletContextTemplateResolver;
-
-@Configuration
-public class ThymeleafConfig {
-    
-    @Bean 
-    public ServletContextTemplateResolver templateResolver() {
-        ServletContextTemplateResolver resolver = new ServletContextTemplateResolver(null);
-        resolver.setPrefix("/WEB-INF/jsp/");
-        resolver.setSuffix(".jsp");
-        resolver.setOrder(1);
-        return resolver;
-    }
-
-    @Bean 
-    public SpringTemplateEngine templateEngine() {
-        SpringTemplateEngine engine = new SpringTemplateEngine();
-        engine.setTemplateResolver(templateResolver());
-        return engine;
-    }
-
-    @Bean 
-    public ThymeleafViewResolver thymeleafViewResolver() {
-        ThymeleafViewResolver resolver = new ThymeleafViewResolver();
-        resolver.setTemplateEngine(templateEngine());
-        return resolver;
-    }
-}
+//package file.config;
+//
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import org.thymeleaf.spring5.SpringTemplateEngine;
+//import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
+//import org.thymeleaf.spring5.view.ThymeleafViewResolver;
+//import org.thymeleaf.templateresolver.ITemplateResolver;
+//import org.thymeleaf.templateresolver.ServletContextTemplateResolver;
+//
+//@Configuration
+//public class ThymeleafConfig {
+//    
+//    
+//}
