+++ src/main/java/file/config/WebMvcConfig.java
@@ -1,19 +1,81 @@
 package file.config;
 
+import org.apache.commons.dbcp2.BasicDataSource;
+import org.apache.commons.dbcp2.DataSourceConnectionFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.ViewResolverRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+import org.springframework.web.servlet.view.JstlView;
+import org.thymeleaf.extras.java8time.dialect.Java8TimeDialect;
+import org.thymeleaf.extras.springsecurity5.dialect.SpringSecurityDialect;
+import org.thymeleaf.spring5.SpringTemplateEngine;
+import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
+import org.thymeleaf.spring5.view.ThymeleafViewResolver;
+import org.thymeleaf.templateresolver.ITemplateResolver;
 
 @Configuration
-@EnableWebMvc
-public class WebMvcConfig implements WebMvcConfigurer{
-    
+@EnableTransactionManagement
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    // @Override
+    // public void addViewControllers(ViewControllerRegistry registry) {
+    // registry.addRedirectViewController("/logout", "/");
+    // registry.addViewController("/error").setViewName("error");
+    // registry.addViewController("/login").setViewName("login");
+    // registry.addViewController("/").setViewName("index");
+    // }
+
     @Override
-    public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addRedirectViewController("/logout", "/");
-        registry.addViewController("/error").setViewName("error");
-        registry.addViewController("/login").setViewName("login");
-        registry.addViewController("/").setViewName("index");
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/resources**").addResourceLocations("/resources");
+        WebMvcConfigurer.super.addResourceHandlers(registry);
+    }
+
+    // @Bean
+    // public ViewResolver viewResolver() {
+    // InternalResourceViewResolver resolver = new InternalResourceViewResolver();
+    // resolver.setViewClass(JstlView.class);
+    // resolver.setPrefix("/WEB-INF/jsp/");
+    // resolver.setSuffix(".jsp");
+    // resolver.setOrder(1);
+    // return resolver;
+    // }
+
+    @Bean
+    public ITemplateResolver templateResolver() {
+        SpringResourceTemplateResolver resolver = new SpringResourceTemplateResolver();
+        resolver.setPrefix("/WEB-INF/jsp/");
+        resolver.setSuffix(".jsp");
+        return resolver;
+    }
+
+    @Bean
+    public SpringTemplateEngine templateEngine() {
+        SpringTemplateEngine engine = new SpringTemplateEngine();
+        engine.setTemplateResolver(templateResolver());
+        return engine;
+    }
+
+    @Bean
+    public ThymeleafViewResolver thymeleafViewResolver() {
+        ThymeleafViewResolver resolver = new ThymeleafViewResolver();
+        resolver.setTemplateEngine(templateEngine());
+        return resolver;
+    }
+
+    @Autowired
+    public void addDialect(final SpringTemplateEngine engine) {
+        engine.addDialect(new SpringSecurityDialect());
+        engine.addDialect(new Java8TimeDialect());
     }
 }
