+++ src/main/resources/application.properties
@@ -3,3 +3,7 @@ jdbc.url=jdbc:derby:.derby;create=true
 
 spring.liquibase.changelog = classpath:databaseChangelog.xml
 liquibase.contexts=derby
+
+spring.resources.add-mappings=true
+spring.datasource.url=jdbc:derby:.derby;create=true
+spring.datasource.driver-class-name: org.apache.derby.jdbc.EmbeddedDriver
