+++ src/main/webapp/WEB-INF/jsp/error.jsp
@@ -1,6 +1,6 @@
-<%@taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
+<!--%@taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%-->
 
 <html>
 <head>
