+++ src/main/webapp/WEB-INF/jsp/sign_up.jsp
@@ -1,63 +1,62 @@
-<%@taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
+<!--%@taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
 <%@page contentType="text/html;charset=UTF-8" language="java"%>
-<%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+<%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%-->
 
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<html lang="en" xmlns="http://www.w3.org/1999/xhtml" 
+    xmlns:th="http://www.thymeleaf.org"
+    xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity5">
 
-	<head>
+    <head>
 
-		<title>SignUp</title>
+        <title>SignUp</title>
 
-		<link href="<c:url value="/resources/styles/sign_up.css" />" rel="stylesheet">
+        <link th:href="@{/resources/styles/sign_up.css}" rel="stylesheet">
 
-	</head>
+    </head>
 
-	<body>
-		<div class="header">
-			<h2>Sign up</h2>
-		</div>
-		<form:form id="regForm" modelAttribute="userJSP" action="registerProcess"
-			method="post">
-			<div class="login-table">
+    <body>
+        <div class="header">
+            <h2>Sign up</h2>
+        </div>
+        <form id="regForm" th:modelAttribute="userJSP" th:action="@{registerProcess}"
+            method="post" th:object="${userJSP}">
+            <div class="login-table">
+                <table>
+                    <div class="Mail-Validation">
+                        <th:errors path="mail" cssClass="error" />
+                    </div>
+                    <div class="Mail">
+                        <label for="mail">Mail</label>
+                        <input type="text" name="mail" id="mail" 
+                            th:value="${userJSP} ? ${userJSP.mail}"/>
+                    </div>
 
-				<table>
-					<div class="Mail-Validation">
-						<form:errors path="mail" cssClass="error" />
-					</div>
-					<div class="Mail">
-						<form:label path="mail">Mail</form:label>
-						<form:input  path="mail" name="mail"
-							id="mail" />
-					</div>
+                    <div class="Username-Validation">
+                        <th:errors path="name" cssClass="error" />
+                    </div>
 
-					<div class="Username-Validation">
-						<form:errors path="name" cssClass="error" />
-					</div>
+                    <div class="Username">
+                        <label for="name">Your Name</label>
+                        <input type="text" name="name" id="name" 
+                            th:value="${userJSP} ? ${userJSP.name}"/>
+                    </div>
 
-					<div class="Username">
-						<form:label path="name">Your Name</form:label>
-						<form:input path="name" name="name" id="name" />
-					</div>
+                    <div class="Password-Validation">
+                        <th:errors path="password" cssClass="error" />
+                    </div>
 
+                    <div class="Password">
+                        <label for="password">Password</label>
+                        <input type="password" name="password" id="password"/>
+                    </div>
 
-					<div class="Password-Validation">
-					<form:errors path="password" cssClass="error" />
-					</div>
+                    <div class="Register">
+                        <button id="register" name="register">Register</button>
+                    </div>
+                    
+                    </table>
+                </div>
+            </form>
+        </body>
 
-					<div class="Password">
-
-						<form:label path="password">Password</form:label>
-						<form:password path="password" name="password"
-							id="password" />
-					</div>
-
-					<div class="Register">
-						<form:button id="register" name="register">Register</form:button>
-					</div>
-
-					</table>
-				</div>
-			</form:form>
-		</body>
-
-	</html>
+    </html>
